added all source files
These were developed in a private "dev" repo. Many things
were completely scrapped, either for different projects
or just bad ideas in the end. Nevertheless, much
"legacy" code will still be present here in the sense
that the final set of functions necessary for the
QMSC paper data is a smaller subset. However,
there are still useful ideas/ functinos that
we didn't use, so we decided not to try and meticulously
cull things.