added clarity to script file output
Namely, we added a column which specifies which state
was run for the given ensemble.

We also fixed a minor bug with the creation and submission
script "create_jobs_and_run" where things were still labelled
as "vcca" instead of "ccps" which caused the slurm submission
files to not actually have an execution statement running
the python simulation.